requirements.txt
amqp==5.0.5
asgiref==3.2.10
autopep8==1.5.5
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
cryptography==3.4.6
defusedxml==0.6.0
diff-match-patch==20200713
dill==0.3.3
Django==3.1.1
django-debug-toolbar==3.2
django-import-export==2.3.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
drf-yasg==1.17.1
et-xmlfile==1.0.1
google-auth==1.27.1
gunicorn==20.0.4
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.4
packaging==20.4
prompt-toolkit==3.0.16
psycopg2==2.8.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyfunctional==1.4.2
Pygments==2.8.0
PyJWT==1.7.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-utils==2.4.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0